app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
alembic
asyncpg
greenlet
passlib "bcrypt==4.0.1"
PyJWT
python-multipart
python-dotenv